First commit after changing to Iphone notes clone

+++ lib/views/notes/create_update_note_view.dart
@@ -6,6 +6,7 @@ import 'package:mynotes/utilities/dialogs/cannot_share_empty_note_dialog.dart';
 import 'package:mynotes/utilities/generics/get_arguments.dart';
 import 'package:mynotes/services/cloud/cloud_note.dart';
 import 'package:mynotes/services/cloud/firebase_cloud_storage.dart';
+import 'package:mynotes/views/notes/notes_view.dart';
 import 'package:share_plus/share_plus.dart';
 
 /*class CreateUpdateNoteView extends StatefulWidget {
@@ -122,11 +123,12 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   CloudNote? _note;
   late final FirebaseCloudStorage _notesService;
   late final TextEditingController _textController;
-
+  late final TextEditingController titletextController;
   @override
   void initState() {
     _notesService = FirebaseCloudStorage();
     _textController = TextEditingController();
+    titletextController = TextEditingController();
     super.initState();
   }
 
@@ -137,22 +139,27 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     }
     final text = _textController.text;
     await _notesService.updateNote(
+      title: titletextController.text,
       documentId: note.documentId,
       text: text,
     );
   }
 
   void _setupTextControllerListener() {
+    titletextController.removeListener(_textControllerListener);
+    titletextController.addListener(_textControllerListener);
     _textController.removeListener(_textControllerListener);
     _textController.addListener(_textControllerListener);
   }
 
+  String userId = "";
   Future<CloudNote> createOrGetExistingNote(BuildContext context) async {
     final widgetNote = context.getArgument<CloudNote>();
 
     if (widgetNote != null) {
       _note = widgetNote;
       _textController.text = widgetNote.text;
+      titletextController.text = widgetNote.title;
       return widgetNote;
     }
 
@@ -161,7 +168,7 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
       return existingNote;
     }
     final currentUser = AuthService.firebase().currentUser!;
-    final userId = currentUser.id;
+    userId = currentUser.id;
     final newNote = await _notesService.createNewNote(ownerUserId: userId);
     _note = newNote;
     return newNote;
@@ -169,7 +176,9 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
 
   void _deleteNoteIfTextIsEmpty() {
     final note = _note;
-    if (_textController.text.isEmpty && note != null) {
+    if (_textController.text.isEmpty &&
+        titletextController.text.isEmpty &&
+        note != null) {
       _notesService.deleteNote(documentId: note.documentId);
     }
   }
@@ -179,6 +188,7 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     final text = _textController.text;
     if (note != null && text.isNotEmpty) {
       await _notesService.updateNote(
+        title: titletextController.text,
         documentId: note.documentId,
         text: text,
       );
@@ -190,6 +200,7 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     _deleteNoteIfTextIsEmpty();
     _saveNoteIfTextNotEmpty();
     _textController.dispose();
+    titletextController.dispose();
     super.dispose();
   }
 
@@ -197,9 +208,6 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(
-          context.loc.note,
-        ),
         actions: [
           IconButton(
             onPressed: () async {
@@ -210,9 +218,17 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                 Share.share(text);
               }
             },
-            icon: const Icon(Icons.share),
+            icon: Icon(Icons.ios_share),
           ),
         ],
+        leading: IconButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            icon: Icon(
+              Icons.arrow_back_ios,
+              color: Colors.yellowAccent[700],
+            )),
       ),
       body: FutureBuilder(
         future: createOrGetExistingNote(context),
@@ -220,13 +236,24 @@ class _CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
           switch (snapshot.connectionState) {
             case ConnectionState.done:
               _setupTextControllerListener();
-              return TextField(
-                controller: _textController,
-                keyboardType: TextInputType.multiline,
-                maxLines: null,
-                decoration: InputDecoration(
-                    hintText: "Start typing your note",
-                    border: InputBorder.none),
+              print("bro: ${titletextController.text}");
+              return Padding(
+                padding: const EdgeInsets.all(15.0),
+                child: Column(
+                  children: [
+                    TextField(
+                        controller: titletextController,
+                        maxLines: 1,
+                        style: TextStyle(
+                            fontSize: 30, fontWeight: FontWeight.bold),
+                        decoration: InputDecoration(border: InputBorder.none)),
+                    TextField(
+                        controller: _textController,
+                        keyboardType: TextInputType.multiline,
+                        maxLines: null,
+                        decoration: InputDecoration(border: InputBorder.none)),
+                  ],
+                ),
               );
             default:
               return const CircularProgressIndicator();
